+++ src/core/TuyaOpenAPI.ts
@@ -4,6 +4,7 @@
 import https from 'https';
 import Crypto from 'crypto';
 import { v4 as uuidv4 } from 'uuid';
+import retry from 'async-await-retry';
 
 // eslint-disable-next-line
 // @ts-ignore
@@ -284,7 +285,7 @@ export default class TuyaOpenAPI {
       path += '?' + new URLSearchParams(params).toString();
     }
 
-    const res: TuyaOpenAPIResponse = await new Promise((resolve, reject) => {
+    const res: TuyaOpenAPIResponse = await retry(async () => new Promise((resolve, reject) => {
 
       const req = https.request({
         host: new URL(this.endpoint).host,
@@ -310,9 +311,12 @@ export default class TuyaOpenAPI {
         req.write(JSON.stringify(body));
       }
 
-      req.on('error', e => reject(e));
+      req.on('error', e => {
+        this.log.error('Netrork error: %s. Retrying...', e.message);
+        reject(e);
+      });
       req.end();
-    });
+    }), undefined, {retriesMax: 10, interval: 100, exponential: true, factor: 2, jitter: 100});
 
     this.log.debug('Response:\npath = %s\ndata = %s', path, JSON.stringify(res, null, 2));
     if (res && res.success !== true && API_ERROR_MESSAGES[res.code]) {