+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 - Add option to make accessory and unbridged accessory (#285) Thanks @donavanbecker for the contribution
 - Add inching button for switches.
 - Add support to 2ch windows covering. (#339) Thanks @CryptoIR for the contribution
+- Add retry when network error happend.
 
 
 ### Fixed